apollo_consensus: extract consensus clock into `apollo_time`
Only EXTRACT, no logic changes.

TODO(gilad): use everywhere (will be implemented after we finish the
milestone).
Also add `FakeClock` instead of mockall: impl is similar to
the mempool's clock, which will also be converted to use this clock
rather than instants.
