+++ crates/apollo_time/Cargo.toml
@@ -5,10 +5,18 @@ edition.workspace = true
 repository.workspace = true
 license.workspace = true
 
+[features]
+testing = ["mockall"]
 
 [dependencies]
+async-trait.workspace = true
+chrono.workspace = true
+mockall = { workspace = true, optional = true }
+tokio = { workspace = true, features = ["time"] }
+
 
 [dev-dependencies]
+mockall.workspace = true
 
 [lints]
 workspace = true
